ci: adds logs, extra checks to Cocoapods DemoApp test

+++ scripts/run-tests-cocoapods-monterey.sh
@@ -1,15 +1,33 @@
 #!/bin/bash
 set -euo pipefail
 
-echo "Running unit tests on Xcode version: $(xcode-select -p)"
+readonly XCODE=$(xcodebuild -version | grep Xcode | cut -d " " -f2)
 
-# Delete the old stuff
+if ! command -v xcbeautify &> /dev/null
+then
+    echo -e "\033[1;31m ERROR: xcbeautify could not be found please install it... \033[0m"
+    exit 1
+fi
+
+echo "▸ Current Xcode: $(xcode-select -p)"
+
+echo "▸ Using Xcode Version: ${XCODE}"
+
+echo "▸ Available Xcode SDKs"
+xcodebuild -showsdks
+
+echo "▸ Removing XCFramework folder"
 rm -Rf XCFramework
-# reset simulators
+
+echo "▸ Shutdown all simulators"
 xcrun -v simctl shutdown all
+
+echo "▸ Erase all simulators"
 xcrun -v simctl erase all
 
+echo "▸ Unzipping XCFramework"
 unzip MUXSDKStats.xcframework.zip
+
 cd apps/DemoApp
 
 echo "▸ Reset Local Cocoapod Cache"
